pyproject.toml
[project]
name = "andreabonn-readme"
version = "1.0.0"
description = "GitHub profile README automation — basket-themed widgets"
requires-python = ">=3.11"
dependencies = [
    "Pillow>=10.0,<13",
    "requests>=2.31,<3",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501", "SIM108"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["scripts"]

[tool.coverage.run]
source = ["scripts"]
omit = ["scripts/common/__init__.py"]

[tool.coverage.report]
fail_under = 85
show_missing = true

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "ruff>=0.4",
    "bandit>=1.9.4",
    "pip-audit>=2.10.0",
]
